requirements.txt
asgiref==3.8.1
astroid==3.1.0
colorama==0.4.6
dill==0.3.8
Django==5.0.4
django-taggit==5.0.1
isort==5.13.2
mccabe==0.7.0
pillow==10.3.0
platformdirs==4.2.1
psycopg2-binary==2.9.9
pylint==3.1.0
sqlparse==0.5.0
tomlkit==0.12.4
tzdata==2024.1
